Added Generic multichannel frequency converter
Generic Frequency converter for use with Voltage to Frequency converters. The IP Core counts the pulses comming in to a port for a specified time window. The value of count is storaged into a internal register. For example, if the input frequency at C_I(0) is 10KHz and the window time is setup to 100ms = 0.1s then, the register at ADR_I=0 will have 10,000 Hz / 0.1 s = 1,000 counts.