Added EASYDRV v1.0
Description: SBA Simple step motor adapter for Easy Driver and similar
hardware. The IP Core have a register where the current position (currpos) is
storage, this register can be reset internally using the corresponding bit
in the control register or externally using the input port RSTPOS.
Writing to the set position register (setpos) instruct to the adapter to
source the appropriate signals trough the DIR and STEP outputs to achieve the
new position. When the step motor arrive to the destiny position a flag is
set to '1' in the status register. The IP Core controls the STEP speed and
acceleration.